requirements.txt
###### Requirements without Version Specifiers ######`
python-telegram-bot
opencv-python
numpy
Pillow